Daemonization bugfixes
This fixes a number of issues with gdnsd's
daemonization code (pidfile handling, and
self-init-script stuff).  This should prevent
problems with close timing races on start/restart,
and also fix issues with not reporting correct
initscript exit values for those operations
under some failure modes.

The new code maintains an fcntl F_WRLCK lock
on the pidfile for the life of the daemon, which
is obtained after the final daemonization fork,
and communicates daemonization success back to
the original invoking process via pipe() to
provide proper exit values.

The actual daemon pid is still written to the
pidfile as text for informational purposes,
but the actual gdnsd init commands (such as
status, stop, etc) use the active fcntl lock to
obtain the pid of the running daemon, which should
prevent any issues with killing unrelated
processes due to stale pidfile info and pid
reuse, etc.