Backwards-compat for pidfile stuff
This should enable correct functionality when the 1.6.8
code takes action while a pre-1.6.8 daemon is still running,
allowing for smoother upgrades.  It tries to read the
pid from unlocked pidfiles for the pid number, and
trusts it if any process is running at the stated pid.

This code won't be merged to 1.7.x and beyond since easy
upgrades will be broken there regardless and these compat
fixes enable some corner-case bad behaviors.