ztree-level duplicate zone sorting
If more than one data source for a given zone is alive at one
moment in time (virtual time as measured by the serial
invocations of ztree_update()), they will now always be sorted.

The sort criteria is first by the zone serial number from the
SOA record, and then (if serials are equal) by the mtime
provided by the source (file mtime on disk in the rfc1035
case).  If both serial and mtime compare equally, the
most-recently updated file in terms of the realtime
ztree_update() invocation timeline wins.

Note: serial math is naive in this first version, I need to
look up how to do it properly to handle wrapping correctly.