ztree lock/sort cleanup and transaction mode
This probably fixes a few bugs with the ztree locking
logic and the duplicate sorting.  It also enables an
API for batching several zone data updates into a single
transaction (not yet tested or used by other code, but
I think I'll had a SIGHUP hook to rfc1035 to load all
disk changes as one txn, might be a nice alternative for
those that don't want constant automatic scan+reload).