SIGHUP work and "zreload" config params renamed
SIGHUP now triggers a zone data reload, and the "reload"
commandline/initscript action was changed to send this SIGHUP
rather than doing a full daemon restart.

The parameters for enabling/disabling/tweaking automatic reloads
of rfc1035 zone data were renamed and reworded.  Gave up on the
idea of transactional reload for the rfc1035 case.  It really
doesn't make a ton of sense in the net (since there's no
transactional feedback to the entity that sent SIGHUP after
several modifications to data), and it started looking like
a real PITA to implement.

Related docs updates.