fix/refactor on quiesence timers
Basically we can't avoid doing a quiescence wait
on daemon startup, or else we can't reliably detect
changes that happen an instant after we initially
read the file (but too close for an mtime update).
They'd simply never get picked up no matter how
many times you HUP'd the daemon, unless the user
touched the altered file again.

The portable 1.02s minimum quiesce wait also sucks
to go through at startup every time during a testsuite
run (although it's really not a big deal for normal
usage), as it would add a ton of time to the testsuite
overall.

For the sake of general-case speed, I've added some
heuristic code to take advantage of st_mtim.tv_nsec
and adjust the minimum quiesce wait down to 10ms if
the zones filesystem appears to support this usefully.

For the sake of the testsuite, I've added an undocumented
env var that the testsuite can use to force skipping
the startup delay when we know we're not planning to
do any runtime zonefile mods during a specific test
run, and enabled it by default for all tests.  It
just forces *all* quiesce timers down to 0.0.

The actual tests of runtime zonefile modifications
on low-res filesystems (e.g. the common Linux ext3)
are just going to run a bit slow and that can't be
helped if the tests are to be reliable.  ext4, btrfs,
jfs, tmpfs, and I believe xfs should all support high
res timestamps if you care enough to switch.