Update plugin-api docs re: full_config(), etc...
In the 1.6.x runtime model, configuration was always loaded even
for simple actions like "stop" (because the pidfile location was
in the config file).  Thus the plugin callback full_config() was
split off as something only run when the daemon was really starting
up, for plugins to do expensive post-config setup.

In the modern runtime model, the simple actions like "stop" do
not load the daemon configuration at all (and thus also do not
load plugins at all), and so heavy lifting can be performed
immediately in the earlier plugin config actions.  full_config()
still exists for backwards compat and because there could still
be reasons to use it for late setup of per-iothread structures, etc,
but full_config() always runs immediately after the other config
related callbacks, and there's no advantage to shifting heavy
operations out to it just for the sake of faster execution in
the earlier phases.