Further systemd service config fixup.
This copy basically functions, but has a number of caveats.  At
this point I'm just tired of dealing with systemd's crap, and I
guess users who can't use sysv-style initscripts get what's coming
to them: a poorer experience with gdnsd.

Details of the various frustration points are now documented in
the servicefile