Explicitly link liburcu-qsbr/libev with programs
Binaries use liburcu-qsbr's & libev's symbols directly and not via
libgdns, the former via gdnsd-prcu-priv.h which is just preprocessor
defines, the latter via direct calls to libev.

Add those two libraries to LDADD for all the non-plugin binaries. It
works as it is because libtool expands the dependency automatically.
That particular libtool feature (link_all_deplibs) is, however, disabled
in Debian/Ubuntu's, which causes gdnsd to fail to build if libtoolize is
first run on the tree on such systems.

Technically, the same should be added to all plugins, as they also use
the symbols directly (or, alternatively, prcu-priv can be made part of
libgdnsd's ABI). However, they're correctly exposed from the main gdnsd
binary and it build properly now, so keep it unchanged in the interest
of laziness :)