Install development headers under pkgincludedir
Install the headers under pkgincludedir, typically /usr/include/gdnsd
instead of the /usr/include.

Unfortunately, this leads to /usr/include/gdnsd/gdnsd-foo.h headers.
Ideally, these should be changed to gdnsd/foo.h but this would be a
massive change as it affects all in-tree users (which is quite a lot).