Provide compatibility wrappers for {recv,send}mmsg
recvmmsg() is supported since 2.6.32 and glibc 2.12 and sendmmsg() is
supported since 3.0.0 and glibc 2.14. There are distributions out there
that run a sufficient enough kernel but don't have a new enough glibc,
as a) there was a significant time interval between kernel and libc
support, b) it's easier to upgrade a kernel than the libc on a running
system.

For example, Debian squeeze has glibc 2.11 but can run a (backported)
3.2.0 kernel just fine. Even worse, Debian's *upcoming* version will
release with 3.2.0 and glibc 2.13 and that will remain "stable" for
about two years.

For such cases, provide alternative versions for recvmmsg() and
sendmmsg() based on syscall definitions from the Linux headers.