Fix libgdmaps for system paths
Commit 747fce made chroot optional and switched to system paths. This
fixes the gdnsd_geoip_test binary to also work with such a scheme (see
issue #23).

This shamelessly copies conf_load_vscf() from conf.c (replacing
conf_load() which was also copied as far as I can see). It might make
sense for conf_load_vscf() to be moved and shared across the two but for
now this will do.