Fix ztree_txn_* API to work
ztree_clone() sets ztclone->zones to valid pointer even where
it was NULL in the original ztree. This happens since malloc(0)
returns always valid pointer. This confuses _ztree_update() and
various other places as in several cases ->zones is tested
instead of ->zones_len.

Fix ztree_clone() to keep ->zones as NULL if the original had
it as NULL.