Remove additional-section processing of PTR records
The standards are pretty clear on this, but I somehow
missed it in the past.  It's not a practical consideration
in any real scenario, since useful PTR records invariably
have the right-hand-side canonical name in a different zone
than the one the PTR record existed in.  The support for
in-zone additional lookups was just there "for completeness",
but it turns out to not be necessary at all.

I think anyone affected by this change is using PTR records for
something crazy that they're *really* not intended to do, and
even in that case it's not going to break anything, just cause
an additional round-trip for the data.

The old warning that happened during the additional-section
processing of PTRs was useful for catching reverse-zonefile typos,
though, and it has been retained in spirit, it's just implemented
differently.