Bugfix for geoip 'nets' file normalization
This is not the first time that we've had a bug
here.  The code is just too complex trying to do
it all in one fast pass with a bi-directional cursor
over the data.  For now, I've switched to a much
simpler model that iterates forward-only normalization
passes until no further normalizations are found.

The new code is slower, but at least it's much more
likely to be correct.  We can always go back and find
other ways to optimize if it becomes an issue for someone.

Issue #62