Put back up/down builtins and defaulting to up when unspecified
This is a partial revert of the changes in 4a602b77 - ultimately
the design tradeoffs just didn't work out as I'd hoped.  What's
been put back is just the existence of fixed "up" and "down"
as built-in service types that don't have to be defined, and
having plugins default to "up" if no service_types is set at all.

However, 4a602b77's other improvements remain, including that if
you explicitly define service_types => [], not even the default
"up" service_type is used, and thus there is no monitored-status
output or ability to administratively override anything (as well
as the sttl_updater, plugin_static stuff, and general compatibility
of sttl_min() with zero-length lists for [] support).