New plugin_extfile
This is similar to admin_state (same file format, anyways), but
tailored for the mass importation of monitoring results from other
sources via an external textfile.  Seems prudent to add such a
thing at this stage, or else some will be inclined to abuse the
admin_state file for this.