admin state user/tool input stuff
Basically, the file /var/lib/gdnsd/admin_state is watched with
an ev_stat watcher.  When it's present, it's parsed for vscf hash
entries like:
   127.0.0.1/www_svc => DOWN

The left-hand-side are actually glob patterns, and a TTL can optionally
be forced as well (subject to zonefile range clamping), e.g.
   127.0.0.1/* => UP/300

What's being matched on the left-hand-side is the "description", which
for true monitors is "addr_or_cname/service_type", and for virtuals is
plugin-defined (e.g. "geoip/res4/dc5").

An empty file is the same as one that's not present.  If an active file
is deleted, all forced states are cleared.