dynamic address limits for resolver plugins
Previously, the maximum number of dynamic addresses that
could be returned by a plugin in a result-set was fixed
to a hardcoded limit of 64x IPv4 + 64x IPv6.

With this change, the limit becomes dynamic with respect
to user configuration.  Plugins are responsible for notifying
the core of the maximum results they can return at config-time
(by examining the user's config data).  The new internal
minimums also consume considerably less space than the old,
fixed, storage sizes of 64+64.

As a part of this change, dyn_result_t has become a private
structure which is only updateable by plugins through exported
function interfaces.

Note that we still enforce an absolute maximum of 512 addresses
of each family in the current code as a sanity limit.  Removing
or raising this limit won't require plugin-API changes, so it
could be done in a minor release in the future relatively quickly.
There may be other internal issues to resolve as a part of such
a change, however, aside from just updating/removing the limit
check itself.

cf. https://groups.google.com/forum/?fromgroups=#!topic/gdnsd/vdAvCPyYhTE