Remove mon_loop ev_child watcher in plugin_extmon
It was always wrong to do this here, it just happened to
work because mon_loop == def_loop (which will shortly no
longer be the case!) and no other plugins were also spawning
and watching children.

We're catching and logging the death in the pipe reader anyways,
so there's no true loss of reporting.  We will gain a zombie
now on these deaths, but that's ok - it's only ever one
process per daemon, and if anything it further highlights
the problem.  These children should never die unless there's a
code bug involved, and we never try to respawn them if they do
(because we can't in the chroot case).