Remove loop argument from plugins' pre_run callback
It was never wise to expose the main thread's default
ev loop this way.  No plugin is currently using this facility,
and the default loop of the main thread is going away in the
near future.  Plugins that really need to do complex
maintenance tasks can spawn their own side-threads the
way that plugin_geoip does.