docs autotools refactor + simple path templating
Re-centralized the docs to make autotools stuff simpler

Got rid of pod->txt for gdnsd_manual, can just be written
in text.

Stopped installing pod/txt copies of the installed manpages.

Added a sed-based system for templating the configured
default filesystem paths into the manpages, so the docs
can be clearer.  The templated directory variables look like:
@GDNSD_DEFPATH_RUN@

Made perl + pod2man required for build so that we don't
have to ship intermediate outputs for the new templating
in tarballs and so that the automake rules are simpler.