Change "-c" commandline param to config dir, not file.
This implies a lot of other internal changes and doc changes for
the better, and is an improvement on the initial chroot-removal
work in ae603f03.  Switching to specifying the config *directory*
instead of file simplifies things as there are no longer
multiple layers of default to explain, and we no longer have to
speak about "the directory containing the current config file"
in the docs, either.

The situation now is that there is a default config directory,
which can be overridden on the commandline.  The primary config
file is the file 'config' within that directory.  Other configuration
(e.g. zones/ subdir, geoip/ subdir) live within that directory
as well.