New config option 'max_edns_response'
Some clients apparently default to advertising a 4K edns0
buffer size for UDP response traffic, which isn't well-suited
to use on the public Internet.  Some rare datasets may actually
have data this large for irregular usage of the DNS to transport
other data.  The intersection of these two things can cause
problems.

This option caps the buffer size advertised by the client to
a configurable value on the server, so that we don't try to
send edns0 UDP responses larger than this value, instead
trucating and falling back to TCP.  It defaults to a sane
value of 1280, which is very unlikely to fragment on the public
Internet, yet still offers more than twice the packet size
of traditional UDP DNS, and more than enough for normal,
non-DNSSEC responses in most cases.

See discussion here:
https://groups.google.com/forum/?fromgroups=#!topic/gdnsd/S6CuZj18nVQ