Fix sizing issues with max_edns_response, etc
Obviously, for the testsuite to do its job max_edns_response
needs to have the same upper limit as max_response, and
needs to be set for some tests.

Along the way it was noticed that the docs and code disagreed
on the upper limit of max_response.  64000 has been the correct
and functional value for a long time, and the docs have been
incorrectly stating that it was 62464.  The limit was raised
from 62464 to 64000 back when the DNSCurve code was removed
prior to the 1.6.0 release.