Change down_thresh min value from 1 to 2.
Minimum value of 1 leads to broken behavior where
a resource would never enter the DOWN state.  There
are multiple ways this could be resolve, but this one
avoids behavior impact on other cases people might
actually be relying on in 1.x.

(Also clears up outdated comments and docs on the
max value of the thresholds).