s/readdir()/readdir_r()/
Better to be safe than sorry.  Currently, the only users
of readdir() in the codebase are both in the same thread
which makes this a non-issue in practice, but there's no
guarantee that will always be the case, or that 3rd-party
plugins won't want to read directories in their own threads.