vscf: s/const vscf_data_t*/vscf_data_t*/ for pub interface
The original idea was that only the vscf-handling code itself
should be messing with vscf_data_t, therefore its interfaces
which returned that type always returned it as a pointer-to-const,
even though it needed to be modified by later vscf calls.  This
had the obvious downside of introducing unnecessary pointer casts
between const and non-const both within and without the vscf code
itself.

Later the data type became completely opaque to consumers anyways,
making the const-protection largely pointless.  This commit removes
the const-protection and thus gets rid of the ugly const-casting.