Drop UDP source port zero as UDP recvfail stat
Currently on Linux we sometimes receive packets with a UDP
source port of zero.  We process responses to these, and
the kernel refuses to send the reply due to the port zero
destination, so we end up syslogging a failure (per-packet!)
and counting them against our UDP sendfail stat.

This avoids the syslog entry and counts them as recv
failures instead.