kill systemd hax
We still have very basic systemd support, but not the crazy
stuff for hacking cgroups, etc.

I'm not going down this road, especially given that even the
documented behaviors of systemd are such a moving target.
Those who care about the fast restart mode will have to not
run under systemd for now.  I have a better, generic
solution brewing that will allow fast-restart in a clean
manner regardless of systemd, but I'm not sure when it will
be ready.