s/reload/reload-zones/, remove force-reload
This further extends the work in e3f56430 to avoid
compatibility breakage in the future when the new
reload feature is merged post-2.0.0.  It replaces
the action "reload" with "reload-zones" and removes
the action "force-reload".

The new reload code is a feature-in-progress, but
I don't want to delay 2.0.0 any further waiting on it.
With these changes, it should be possible to slip it
into a 2.N.0 release relatively painlessly.