Remove threadnum from plugin resolve() callback args
... and related fallout.  Basically now that TLS is
available, this argument and its storage in
dnspacket_context_t is superfluous.

For plugins which really need access to the thread number at
resolve() time (although honestly, TLS should suffice on its
own), a simple solution is to store the threadnum given in
iothread_init() into a TLS var and access it from resolve().