Switch context/stats pointers from TLS back to args
This undoes some of the work in ff2deb5c (although it retains the
context/stats split), and essentially all of the work in f4b12fc3.
I don't know what I was thinking during that TLS frenzy, but it's
not a net win to switch to TLS in a scenario like this, even it
does make the code prettier.