Document log_stats option; restore "log stats at shutdown" feature
Apparently the log_stats option was never in the documentation due
to some oversight long ago (and its configurable range was strange
and had little to do with reality).  This documents it as being
settable in the range 0-86400, defaulting to 3600, with zero
disabling the periodic log output.

This patch also restores the behavior that gdnsd outputs its final
stats counters to syslog at daemon shutdown time.  This feature
was removed during some threads refactoring a while back, but I
think it's better to keep it even if it's a little trickier to
implement with how the threads work now.