Always generate gdnsd.service
Some packaging systems that are systemd-aware might deal with
things better if "make install" doesn't install the unit file.
This gives them the option to use --without-systemdsystemunitdir
to avoid installation, but still have a generated unit file at a
known path in the build tree to pick up with their own tools.