make scanner setjmps less likely to clobber
gcc 4.9 newly warns about this situation for this code in the
rfc1035 case, and the fix seems pretty straightforward.  Even
with the old code I haven't observed actual misbehavior, though.
Just better to be safe than sorry if it's warning about the
possibility.  Also updated the djb setjmp code to mirror the
current rfc1035 code.