Refactor PRNG code/interfaces in libgdnsd
This bumps the plugin API version to 17 as well and will require
bumping the release version 2.2.0 when the time comes.  This could
have been handled in a backwards-compatible way even at the ABI
level, but (a) beff52ac requires an API version bump + recompile
anyways and (b) the source changes to accomodate this are trivial
for any plugin that needs them.