Support ?f=1 query arg for stats flush-after-fetch deltas
The stats URLs now support a query parameter "f=1" to request
stats flush-after-fetch behavior.  Stats queries that specify f=1
get counter values since the previous f=1 request rather than
since daemon start (unless it's the first such request since
daemon start, obviously).  Stats queries that do not specify the
f=1 argument are unaffected and continue to see the full counts.

Note that the f=1 state is global: if two independent clients are
querying with f=1, they'll reset each others' view of the stats!

Also note that rollover issues in the underlying counts are going
to screw this up pretty bad on 32-bit machines with large amounts
of traffic (but that was true even without this feature.  Use a
64-bit machine if you care!)

This updates the HTTP "parser" a bit as well.  Ultimately I still
think the mini-pseudo-HTTP server should die, but this is a decent
and simple feature to tide some over until it's replaced with
something better.

Issue #76