remove bad/pointless uses of restrict
The ones in the args for dname_from_raw() are pointless since the
compiler can deduce this for a static function (they were brought
over from when it used to be an external one).  The one within
encode_rrs_txt is difficult to understand, and thus we're probably
better off removing it than trying to understand it.