Require ragel for tarball builds (and related refactoring)
Ragel has become a lot more standard and available in recent
years, and this saves us from distributing an extra ~544K of
machine-generated C source code in tarballs.  It also brings the
tarball and git repo forms of the tree into closer alignment.

As of this commit, the diffs in both directions are:
  In the tarball but not the git repo:
    All the files generated by "autoreconf -vi"
  In a git repo checkout but not the tarball:
    /.travis.yml and any file/dir that matches /^\.git/

This commit also incidentally cleans up some Makefile.am nits,
including one that would sometimes break parallel make in the
libgdnsd directory due to a dependency issue.  It also gets rid of
the previous bad habit of generating ragel output into the source
(rather than build) tree.