Use -fvisibility=hidden and gc-sections linker stuff
This replaces our previous use of export-symbols-regex for
controlling symbol visibility in DSOs, and also turns on gcc's
function/data sections with link-time garbage collection to kill
dead code/data.

There was no net change on the visibility front, as we were
already managing that well with a combination of regex filtering
and aggressive use of the static keyword, but this is the better
technique in the long run, especially as things are likely to more
complex down the line.

Some header contents were re-arranged so that the necessary extern
definitions are in a single block for wrapping in a default
visibility pragma.  Those compilers not supporting the gcc-like
visibility flags/pragmas (which arose around gcc 4.0) should still
build the code fine, they'll just potentially be exporting more
symbols than they need to.  I don't think this is a major problem
in practice.