Refactor logf_dname() into gdnsd_dname_to_string(), etc.
logf_dname() was the only dname-to-string formatter available, and
its output was by definition in the log format buffer, making it
unsuitable for general use.  This commit:

(a) Moves the core formatting function to
    gdnsd_dname_to_string() for general use outside of a log
    formatter, which logf_dname() is now a thin wrapper around
    for logging purposes.
(b) Refactors and simplifies the formatting code itself.
(c) Gets rid of the case where zscan_djb was using logf_dname()
    outside of a logging context for lack of a better option.