More signedness-related code cleanups
These are the remainder of cleanups for Wsign-conversion that
seemed like reasonable improvements on the current situation.
There are still several Wsign-conversion warnings in the codebase
in general.

Most of the remaining ones fall under the broad category of
implicit signed int upconversions from char and uint8_t where math
is being done on the numerical value of a byte.  I'm not yet sure
how or if I want to address those.