Fix sendmmsg() error handling
sendmmsg() results other than "all packets sent" are probably
rare to begin with given our buffer sizing and method of operation
here, but the various other cases weren't being handled correctly.
This could have resulted in unecessarily-dropped packets (which
were capable of being retried) and/or meaningless or confusing
error messages when sendmmsg() sent only some of the requested
packets in the first call or had a hard error.

I might have never noticed this if not for looking through the
signedness issues about the return value and staring at the
updated docs for sendmmsg() as a result.

backported from master commit aea98584