Bugfix for extmon timeout/interval >255s
The general monitoring core in gdnsd/libgdnsd/mon.c allows for a
timeout up to 300s and an interval up to 3600s.  However, the
extmon code was truncating these to a single byte for transmission
to the helper over a pipe, which results in the helper using the
raw values modulo 256, which would result in substantially reduced
timeouts and intervals for values >255.

This bug became more-obvious as a result of looking at
type-conversion issues in the network pipe traffic code that
communicates these values.