refactor source tree layout
In general, these steps were taken to reduce the depth of the
source tree for humans and for make recursion, improve parallelism
a little, and reduce the number of generated Makefiles.  Some of
this is also with an eye towards possibly eliminating some or all
of the use of recursive make in the future.

1) libdmn folded into libgdnsd
- They're still conceptually separate, but this avoids the annoying
step of copying the header file and all of the Makefile/tools
fallout that causes.  Lately libdmn has not been tracking very
well towards being a separate library anyways, although hopefully
that will reverse course with future refactorings.

2) moved gdnsd/libgdnsd to top level
- There's no good reason to have this as a recursive subdir anymore,
so may as well flatten the hierarchy a bit.  Will help with
dependency ordering issues later as more unit tests are created.

3) moved plugins/meta/libgdmaps to top level
- This is a separate library as well, so flatten it out to the top

4) renamed gdnsd/ to src/
- This eliminates ambiguities for include paths like <gdnsd/foo.h>
between the libgdnsd "public" headers and the private ones for the
main daemon source, which we've only avoided so far by carefully
avoiding naming conflicts.

5) plugins subdirectories merged up to plugins/
- To reduce pointless Makefile recursion