Bugfix for zero-length strings in vscf
Apparently it's not common in any practical way, but it's
technically legal to have zero-length string values in a vscf
config file, e.g.: foo => ""

Since the allocator changes that were introduced in 2.0.0, our
allocators specifically disallow 0-byte allocations to avoid any
ambiguity about what that means in the standards, and zero-length
strings were generating zero-length allocations.

This fixes it by having the internal raw storage for vscf strings
always be len+1 and setting the final byte to '\0'.  It's simpler
than doing a broader structural fix to support NULL there, and
it's a nice safety valve in case of misuse of the raw value with
functions that expect NUL-termination.