libgdmaps tests: replace "check-download" w/ submodule, etc
This moves the binary testsuite data to a git submodule as
XZ-compressed files.  The autotools setup for this is designed
such that it's completely optional.  You can ignore the submodule
(not initialize it) and things will work fine, you'll just see
that a bunch of unit tests get skipped with messages like "Missing
database".  Ditto for tarball copies, which don't get this data
and don't even have the option of pulling in the submodule.