admin_state: if unparseable, fail startup/checkconf
This is better than current behavior at least.  This and the
zones_strict_startup thing highlight the need in the long run to
develop an interface for transactionally applying new data/state.
Otherwise we always have this disparity between what strictly
fails at startup vs what just logs an error when it fails to
reload at runtime on update.